>chores: Rename automatic tracking wrapper file
The file containing the automatic tracking wrapper was renamed to
reflect the new directory structure.
