+++ birds_nest/pybirdai/process_steps/pybird/automatic_tracking_wrapper.py
